backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
openai==1.61.1
python-dotenv==1.0.1
pydantic[email]==2.9.0
dateparser==1.2.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.146.0
PyPDF2==3.0.1
pandas==2.2.3
apscheduler==3.11.2
python-multipart==0.0.26
pytesseract==0.3.13
Pillow==12.2.0
pytz==2024.2
